add check_mode support to ec2_vpc_net module

+++ cloud/amazon/ec2_vpc_net.py
@@ -144,12 +144,16 @@ def update_vpc_tags(vpc, module, vpc_obj, tags, name):
 
     tags.update({'Name': name})
     try:
-        current_tags = dict((t.name, t.value) for t in vpc.get_all_tags(filters={'resource-id': vpc_obj.id}))
-        if cmp(tags, current_tags):
-            vpc.create_tags(vpc_obj.id, tags)
-            return True
-        else:
-            return False
+        if vpc_obj:
+            current_tags = dict((t.name, t.value) for t in vpc.get_all_tags(filters={'resource-id': vpc_obj.id}))
+            if cmp(tags, current_tags):
+                if not module.check_mode:
+                    vpc.create_tags(vpc_obj.id, tags)
+                else:
+                    vpc_obj.tags = tags
+                return True
+            else:
+                return False
     except Exception, e:
         e_msg=boto_exception(e)
         module.fail_json(msg=e_msg)
@@ -157,8 +161,11 @@ def update_vpc_tags(vpc, module, vpc_obj, tags, name):
 
 def update_dhcp_opts(connection, module, vpc_obj, dhcp_id):
 
-    if vpc_obj.dhcp_options_id != dhcp_id:
-        connection.associate_dhcp_options(dhcp_id, vpc_obj.id)
+    if vpc_obj and vpc_obj.dhcp_options_id != dhcp_id:
+        if not module.check_mode:
+            connection.associate_dhcp_options(dhcp_id, vpc_obj.id)
+        elif dhcp_id:
+            vpc_obj.dhcp_options_id = dhcp_id
         return True
     else:
         return False
@@ -194,6 +201,7 @@ def main():
 
     module = AnsibleModule(
         argument_spec=argument_spec,
+        supports_check_mode=True,
     )
 
     if not HAS_BOTO:
@@ -231,7 +239,8 @@ def main():
 
         if vpc_obj is None:
             try:
-                vpc_obj = connection.create_vpc(cidr_block, instance_tenancy=tenancy)
+                if not module.check_mode:
+                    vpc_obj = connection.create_vpc(cidr_block, instance_tenancy=tenancy)
                 changed = True
             except BotoServerError, e:
                 module.fail_json(msg=e)
@@ -254,15 +263,17 @@ def main():
         # which is needed in order to detect the current status of DNS options. For now we just update
         # the attribute each time and is not used as a changed-factor.
         try:
-            connection.modify_vpc_attribute(vpc_obj.id, enable_dns_support=dns_support)
-            connection.modify_vpc_attribute(vpc_obj.id, enable_dns_hostnames=dns_hostnames)
+            if not module.check_mode:
+                connection.modify_vpc_attribute(vpc_obj.id, enable_dns_support=dns_support)
+                connection.modify_vpc_attribute(vpc_obj.id, enable_dns_hostnames=dns_hostnames)
         except BotoServerError, e:
             e_msg=boto_exception(e)
             module.fail_json(msg=e_msg)
 
         # get the vpc obj again in case it has changed
         try:
-            vpc_obj = connection.get_all_vpcs(vpc_obj.id)[0]
+            if not module.check_mode:
+                vpc_obj = connection.get_all_vpcs(vpc_obj.id)[0]
         except BotoServerError, e:
             e_msg=boto_exception(e)
             module.fail_json(msg=e_msg)
@@ -276,7 +287,8 @@ def main():
 
         if vpc_obj is not None:
             try:
-                connection.delete_vpc(vpc_obj.id)
+                if not module.check_mode:
+                    connection.delete_vpc(vpc_obj.id)
                 vpc_obj = None
                 changed = True
             except BotoServerError, e:
